Panel toolbar bug fixes: clear button on Raw Messages & overlap on Plot Panel
