+++ packages/suite-base/src/components/Autocomplete/Autocomplete.tsx
@@ -14,69 +14,23 @@
 //   found at http://www.apache.org/licenses/LICENSE-2.0
 //   You may not use this file except in compliance with the License.
 
-import {
-  Autocomplete as MuiAutocomplete,
-  Popper,
-  PopperProps,
-  TextField,
-  TextFieldProps,
-} from "@mui/material";
+import { Autocomplete as MuiAutocomplete, Popper, PopperProps, TextField } from "@mui/material";
 import { Fzf, FzfResultItem } from "fzf";
 import * as React from "react";
+import { SyntheticEvent, useCallback, useImperativeHandle, useMemo, useRef, useState } from "react";
+
+import { useStyles } from "@lichtblick/suite-base/components/Autocomplete/Autocomplete.style";
 import {
-  CSSProperties,
-  SyntheticEvent,
-  useCallback,
-  useImperativeHandle,
-  useMemo,
-  useRef,
-  useState,
-} from "react";
-import { makeStyles } from "tss-react/mui";
+  EMPTY_SET,
+  FAST_FIND_ITEM_CUTOFF,
+  MAX_FZF_MATCHES,
+} from "@lichtblick/suite-base/components/Autocomplete/constants";
+import {
+  AutocompleteProps,
+  IAutocomplete,
+} from "@lichtblick/suite-base/components/Autocomplete/types";
 
 import { ListboxAdapterChild, ReactWindowListboxAdapter } from "./ReactWindowListboxAdapter";
-
-const MAX_FZF_MATCHES = 200;
-
-// Above this number of items we fall back to the faster fuzzy find algorithm.
-const FAST_FIND_ITEM_CUTOFF = 1_000;
-
-type AutocompleteProps = {
-  className?: string;
-  disableAutoSelect?: boolean;
-  disabled?: boolean;
-  filterText?: string;
-  hasError?: boolean;
-  inputStyle?: CSSProperties;
-  items: readonly string[];
-  menuStyle?: CSSProperties;
-  minWidth?: number;
-  onBlur?: () => void;
-  onChange?: (event: React.SyntheticEvent, text: string) => void;
-  onSelect: (value: string, autocomplete: IAutocomplete) => void;
-  placeholder?: string;
-  readOnly?: boolean;
-  selectOnFocus?: boolean;
-  sortWhenFiltering?: boolean;
-  value?: string;
-  variant?: TextFieldProps["variant"];
-};
-
-export interface IAutocomplete {
-  setSelectionRange(selectionStart: number, selectionEnd: number): void;
-  focus(): void;
-  blur(): void;
-}
-
-const useStyles = makeStyles()((theme) => ({
-  inputError: {
-    input: {
-      color: theme.palette.error.main,
-    },
-  },
-}));
-
-const EMPTY_SET = new Set<number>();
 
 function itemToFzfResult<T>(item: T): FzfResultItem<T> {
   return {
@@ -204,7 +158,7 @@ export const Autocomplete = React.forwardRef(function Autocomplete(
 
   return (
     <MuiAutocomplete
-      className={className}
+      className={cx(className, classes.root)}
       getOptionLabel={getOptionLabel}
       disableCloseOnSelect
       disabled={disabled}
