+++ packages/suite-base/src/components/PanelToolbar/PanelToolbar.style.ts
@@ -1,0 +1,27 @@
+// SPDX-FileCopyrightText: Copyright (C) 2023-2025 Bayerische Motoren Werke Aktiengesellschaft (BMW AG)<lichtblick@bmwgroup.com>
+// SPDX-License-Identifier: MPL-2.0
+
+import { makeStyles } from "tss-react/mui";
+
+import { PANEL_TOOLBAR_MIN_HEIGHT } from "@lichtblick/suite-base/components/PanelToolbar/constants";
+
+export const useStyles = makeStyles()((theme) => ({
+  root: {
+    transition: "transform 80ms ease-in-out, opacity 80ms ease-in-out",
+    cursor: "auto",
+    flex: "0 0 auto",
+    alignItems: "center",
+    justifyContent: "flex-end",
+    padding: theme.spacing(0.25, 0.75),
+    display: "flex",
+    minHeight: PANEL_TOOLBAR_MIN_HEIGHT,
+    backgroundColor: theme.palette.background.paper,
+    width: "100%",
+    left: 0,
+    zIndex: theme.zIndex.appBar,
+    // eslint-disable-next-line @typescript-eslint/no-explicit-any
+    position: "relative !important" as any,
+    // eslint-disable-next-line @typescript-eslint/no-explicit-any
+    top: "auto !important" as any,
+  },
+}));
