+++ packages/suite-base/src/components/PanelToolbar/index.tsx
@@ -17,52 +17,27 @@
 import FullscreenIcon from "@mui/icons-material/Fullscreen";
 import FullscreenExitIcon from "@mui/icons-material/FullscreenExit";
 import { Typography } from "@mui/material";
-import { useContext, useMemo, CSSProperties } from "react";
-import { makeStyles } from "tss-react/mui";
+import { useContext, useMemo } from "react";
 
 import PanelContext from "@lichtblick/suite-base/components/PanelContext";
+import { useStyles } from "@lichtblick/suite-base/components/PanelToolbar/PanelToolbar.style";
 import ToolbarIconButton from "@lichtblick/suite-base/components/PanelToolbar/ToolbarIconButton";
-import { PANEL_TOOLBAR_MIN_HEIGHT } from "@lichtblick/suite-base/components/PanelToolbar/constants";
+import { PanelToolbarProps } from "@lichtblick/suite-base/components/PanelToolbar/types";
 import { useDefaultPanelTitle } from "@lichtblick/suite-base/providers/PanelStateContextProvider";
 import { PANEL_TITLE_CONFIG_KEY } from "@lichtblick/suite-base/util/layout";
 
 import { PanelToolbarControls } from "./PanelToolbarControls";
 
-type Props = {
-  additionalIcons?: React.ReactNode;
-  backgroundColor?: CSSProperties["backgroundColor"];
-  children?: React.ReactNode;
-  className?: string;
-  isUnknownPanel?: boolean;
-};
-
-const useStyles = makeStyles()((theme) => ({
-  root: {
-    transition: "transform 80ms ease-in-out, opacity 80ms ease-in-out",
-    cursor: "auto",
-    flex: "0 0 auto",
-    alignItems: "center",
-    justifyContent: "flex-end",
-    padding: theme.spacing(0.25, 0.75),
-    display: "flex",
-    minHeight: PANEL_TOOLBAR_MIN_HEIGHT,
-    backgroundColor: theme.palette.background.paper,
-    width: "100%",
-    left: 0,
-    zIndex: theme.zIndex.appBar,
-  },
-}));
-
 // Panel toolbar should be added to any panel that's part of the
 // react-mosaic layout.  It adds a drag handle, remove/replace controls
 // and has a place to add custom controls via it's children property
-export default React.memo<Props>(function PanelToolbar({
+export default React.memo<PanelToolbarProps>(function PanelToolbar({
   additionalIcons,
   backgroundColor,
   children,
   className,
   isUnknownPanel = false,
-}: Props) {
+}: PanelToolbarProps) {
   const { classes, cx } = useStyles();
   const {
     isFullscreen,
