+++ packages/suite-base/src/components/PanelToolbar/types.ts
@@ -1,7 +1,17 @@
 // SPDX-FileCopyrightText: Copyright (C) 2023-2025 Bayerische Motoren Werke Aktiengesellschaft (BMW AG)<lichtblick@bmwgroup.com>
 // SPDX-License-Identifier: MPL-2.0
 
+import { CSSProperties } from "react";
+
 export type PanelToolbarControlsProps = {
   additionalIcons?: React.ReactNode;
   isUnknownPanel: boolean;
+};
+
+export type PanelToolbarProps = {
+  additionalIcons?: React.ReactNode;
+  backgroundColor?: CSSProperties["backgroundColor"];
+  children?: React.ReactNode;
+  className?: string;
+  isUnknownPanel?: boolean;
 };
